Update how screens work
Add onResize handler to the terminal.
When the terminal changes size it closes and reopens
a new screen with the proper size.

Change how menus receive the screen.
They no longer receive the screen itself.
They receive an object of type ScreenGetter.
ScreenGetter gives the screen.
On each redraw a new screen (possibly the same)
is received from the screenGetter.

The main ScreenGetter is the game itself.
When a new screen is Created the game receives
this new screen.
This way it can always give the most recent screen
to the underlying object that needs it.